refactored planning to booking
