+++ src/views/BookingConfirmView.vue
